added context api for the app
context api for this is to allow the user to select their fav flowers and when they decide to plant it, they can click "start planting it" and it will show up in the "planting page". The context api in the case is used to pass down the data from the selected flowers down to "planting" page which make the data handling more efficient and can be reusable.